Initial refactor of HandBrakeCLI-specific bits into its own class
This will allow other ripping backends in time, they only need to implement
the methods that lazybrake calls, and lazybrake won't care about the
internals or which program actually performs the ripping as long as
the end-result (and the returned data) is the same.

This also splits off some common helper methods into its own base class,
lazyBase, so that both the backend and the main logic class can call it.