Added generic dependency checking for backends
This provides a functional usable() method in the base class,
reducing the boilerplate required from a full usable() method to a
simple array of dependencies. This also lets lazybrake display messages
informing the user about why a backend failed.

Also adds a message about which command was not found to the inPath method.